requirements.txt
pyvisa
pyserial
Pyllow
matplotlib
tdqm
scipy
si-prefix